AUTH-1480 - Move email validation to ValidationHelper
- We can avoid the need to instantiate the ValidationService by moving the email validation into static methods. We shouldn't to create an instance of a class to call them.
- Remove a test from CheckUserExistsHandlerTest which was used when the state machine was in the backened api
