+++ frontend-api/src/test/java/uk/gov/di/authentication/frontendapi/lambda/CheckUserExistsHandlerTest.java
@@ -21,7 +21,6 @@ import uk.gov.di.authentication.shared.services.ClientService;
 import uk.gov.di.authentication.shared.services.ClientSessionService;
 import uk.gov.di.authentication.shared.services.ConfigurationService;
 import uk.gov.di.authentication.shared.services.SessionService;
-import uk.gov.di.authentication.shared.services.ValidationService;
 import uk.gov.di.authentication.sharedtest.logging.CaptureLoggingExtension;
 
 import java.util.HashMap;
@@ -49,7 +48,6 @@ class CheckUserExistsHandlerTest {
 
     private final Context context = mock(Context.class);
     private final AuthenticationService authenticationService = mock(AuthenticationService.class);
-    private final ValidationService validationService = mock(ValidationService.class);
     private final AuditService auditService = mock(AuditService.class);
     private final SessionService sessionService = mock(SessionService.class);
     private final ConfigurationService configurationService = mock(ConfigurationService.class);
@@ -66,12 +64,12 @@ class CheckUserExistsHandlerTest {
             new CaptureLoggingExtension(CheckUserExistsHandler.class);
 
     @AfterEach
-    public void tearDown() {
+    void tearDown() {
         assertThat(logging.events(), not(hasItem(withMessageContaining(session.getSessionId()))));
     }
 
     @BeforeEach
-    public void setup() {
+    void setup() {
         when(context.getAwsRequestId()).thenReturn("aws-session-id");
 
         handler =
@@ -81,13 +79,12 @@ class CheckUserExistsHandlerTest {
                         clientSessionService,
                         clientService,
                         authenticationService,
-                        validationService,
                         auditService);
         reset(authenticationService);
     }
 
     @Test
-    public void shouldReturn200IfUserExists() throws JsonProcessingException {
+    void shouldReturn200IfUserExists() throws JsonProcessingException {
         usingValidSession();
         String persistentId = "some-persistent-id-value";
         Map<String, String> headers = new HashMap<>();
@@ -114,16 +111,16 @@ class CheckUserExistsHandlerTest {
                         FrontendAuditableEvent.CHECK_USER_KNOWN_EMAIL,
                         "aws-session-id",
                         session.getSessionId(),
-                        "",
-                        auditService.UNKNOWN,
+                        AuditService.UNKNOWN,
+                        AuditService.UNKNOWN,
                         "joe.bloggs@digital.cabinet-office.gov.uk",
                         "123.123.123.123",
-                        "",
+                        AuditService.UNKNOWN,
                         persistentId);
     }
 
     @Test
-    public void shouldReturn200IfUserDoesNotExist() throws JsonProcessingException {
+    void shouldReturn200IfUserDoesNotExist() throws JsonProcessingException {
         usingValidSession();
         when(authenticationService.userExists(eq("joe.bloggs@digital.cabinet-office.gov.uk")))
                 .thenReturn(false);
@@ -147,8 +144,8 @@ class CheckUserExistsHandlerTest {
                         FrontendAuditableEvent.CHECK_USER_NO_ACCOUNT_WITH_EMAIL,
                         "aws-session-id",
                         session.getSessionId(),
-                        "",
-                        auditService.UNKNOWN,
+                        AuditService.UNKNOWN,
+                        AuditService.UNKNOWN,
                         "joe.bloggs@digital.cabinet-office.gov.uk",
                         "123.123.123.123",
                         AuditService.UNKNOWN,
@@ -156,7 +153,7 @@ class CheckUserExistsHandlerTest {
     }
 
     @Test
-    public void shouldReturn400IfRequestIsMissingEmail() {
+    void shouldReturn400IfRequestIsMissingEmail() {
         usingValidSession();
 
         APIGatewayProxyRequestEvent event = new APIGatewayProxyRequestEvent();
@@ -171,7 +168,7 @@ class CheckUserExistsHandlerTest {
     }
 
     @Test
-    public void shouldReturn400IfRequestIsMissingSessionId() {
+    void shouldReturn400IfRequestIsMissingSessionId() {
         APIGatewayProxyRequestEvent event = new APIGatewayProxyRequestEvent();
         event.setBody("{ \"email\": \"joe.bloggs@digital.cabinet-office.gov.uk\" }");
 
@@ -184,10 +181,8 @@ class CheckUserExistsHandlerTest {
     }
 
     @Test
-    public void shouldReturn400IfEmailAddressIsInvalid() {
+    void shouldReturn400IfEmailAddressIsInvalid() {
         usingValidSession();
-        when(validationService.validateEmailAddress(eq("joe.bloggs")))
-                .thenReturn(Optional.of(ErrorResponse.ERROR_1004));
 
         APIGatewayProxyRequestEvent event = new APIGatewayProxyRequestEvent();
         event.setBody("{ \"email\": \"joe.bloggs\" }");
@@ -203,40 +198,8 @@ class CheckUserExistsHandlerTest {
                         FrontendAuditableEvent.CHECK_USER_INVALID_EMAIL,
                         "aws-session-id",
                         session.getSessionId(),
-                        "",
-                        auditService.UNKNOWN,
-                        "joe.bloggs",
-                        "123.123.123.123",
                         AuditService.UNKNOWN,
-                        PersistentIdHelper.PERSISTENT_ID_UNKNOWN_VALUE);
-    }
-
-    @Test
-    public void shouldReturn200IfUserTransitionsFromUserNotFoundAndUserDoesNotExist()
-            throws JsonProcessingException {
-        usingValidSession();
-        when(authenticationService.userExists(eq("joe.bloggs"))).thenReturn(false);
-
-        APIGatewayProxyRequestEvent event = new APIGatewayProxyRequestEvent();
-        event.setBody("{ \"email\": \"joe.bloggs\" }");
-        event.setHeaders(Map.of("Session-Id", session.getSessionId()));
-        event.setRequestContext(contextWithSourceIp("123.123.123.123"));
-
-        APIGatewayProxyResponseEvent result = handler.handleRequest(event, context);
-
-        assertEquals(200, result.getStatusCode());
-        CheckUserExistsResponse checkUserExistsResponse =
-                objectMapper.readValue(result.getBody(), CheckUserExistsResponse.class);
-        assertEquals("joe.bloggs", checkUserExistsResponse.getEmail());
-        assertFalse(checkUserExistsResponse.doesUserExist());
-
-        verify(auditService)
-                .submitAuditEvent(
-                        FrontendAuditableEvent.CHECK_USER_NO_ACCOUNT_WITH_EMAIL,
-                        "aws-session-id",
-                        session.getSessionId(),
-                        "",
-                        auditService.UNKNOWN,
+                        AuditService.UNKNOWN,
                         "joe.bloggs",
                         "123.123.123.123",
                         AuditService.UNKNOWN,
