+++ shared/src/main/java/uk/gov/di/authentication/shared/helpers/ValidationHelper.java
@@ -5,6 +5,7 @@ import com.google.i18n.phonenumbers.PhoneNumberUtil;
 import uk.gov.di.authentication.shared.entity.ErrorResponse;
 
 import java.util.Optional;
+import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import static com.google.i18n.phonenumbers.PhoneNumberUtil.PhoneNumberType.MOBILE;
@@ -12,6 +13,12 @@ import static com.google.i18n.phonenumbers.PhoneNumberUtil.PhoneNumberType.MOBIL
 public class ValidationHelper {
 
     private static final Pattern PASSWORD_REGEX = Pattern.compile(".*\\d.*");
+    private static final Pattern EMAIL_NOTIFY_REGEX =
+            Pattern.compile("^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~\\-]+@([^.@][^@\\s]+)$");
+    private static final Pattern HOSTNAME_REGEX =
+            Pattern.compile("^(xn|[a-z0-9]+)(-?-[a-z0-9]+)*$", Pattern.CASE_INSENSITIVE);
+    private static final Pattern TLD_PART_REGEX =
+            Pattern.compile("^([a-z]{2,63}|xn--([a-z0-9]+-)*[a-z0-9]+)$", Pattern.CASE_INSENSITIVE);
 
     private ValidationHelper() {}
 
@@ -44,6 +51,45 @@ public class ValidationHelper {
         }
         if (!PASSWORD_REGEX.matcher(password).matches()) {
             return Optional.of(ErrorResponse.ERROR_1007);
+        }
+        return Optional.empty();
+    }
+
+    public static Optional<ErrorResponse> validateEmailAddressUpdate(
+            String existingEmail, String replacementEmail) {
+        if (existingEmail.equals(replacementEmail)) {
+            return Optional.of(ErrorResponse.ERROR_1019);
+        }
+        Optional<ErrorResponse> existingEmailError = validateEmailAddress(existingEmail);
+        if (existingEmailError.isPresent()) {
+            return existingEmailError;
+        }
+        return validateEmailAddress(replacementEmail);
+    }
+
+    public static Optional<ErrorResponse> validateEmailAddress(String email) {
+        if (email == null || email.isBlank()) {
+            return Optional.of(ErrorResponse.ERROR_1003);
+        }
+        Matcher matcher = EMAIL_NOTIFY_REGEX.matcher(email);
+        if (!matcher.matches()) {
+            return Optional.of(ErrorResponse.ERROR_1004);
+        }
+        if (email.contains("..")) {
+            return Optional.of(ErrorResponse.ERROR_1004);
+        }
+        var hostname = matcher.group(1);
+        String[] parts = hostname.split("\\.");
+        if (parts.length < 2) {
+            return Optional.of(ErrorResponse.ERROR_1004);
+        }
+        for (String part : parts) {
+            if (!HOSTNAME_REGEX.matcher(part).matches()) {
+                return Optional.of(ErrorResponse.ERROR_1004);
+            }
+        }
+        if (!TLD_PART_REGEX.matcher(parts[parts.length - 1]).matches()) {
+            return Optional.of(ErrorResponse.ERROR_1004);
         }
         return Optional.empty();
     }
